Se implementan las notaciones  Data y NoArgsConstructor para las clases Entity
