+++ src/main/java/com/levm/expendienteMedico/entity/ExpedienteMedico.java
@@ -3,7 +3,11 @@ package com.levm.expendienteMedico.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@NoArgsConstructor
 @Entity
 public class ExpedienteMedico {
 	
@@ -13,38 +17,10 @@ public class ExpedienteMedico {
 	
 	private String fechaAlta;
 	private String fechaActualizacion;
-	
-	public ExpedienteMedico() {
-		super();
-	}
 
 	public ExpedienteMedico(String fechaAlta, String fechaActualizacion) {
 		super();
 		this.fechaAlta = fechaAlta;
-		this.fechaActualizacion = fechaActualizacion;
-	}
-
-	public long getIdExpediente() {
-		return idExpediente;
-	}
-
-	public void setIdExpediente(long idExpediente) {
-		this.idExpediente = idExpediente;
-	}
-
-	public String getFechaAlta() {
-		return fechaAlta;
-	}
-
-	public void setFechaAlta(String fechaAlta) {
-		this.fechaAlta = fechaAlta;
-	}
-
-	public String getFechaActualizacion() {
-		return fechaActualizacion;
-	}
-
-	public void setFechaActualizacion(String fechaActualizacion) {
 		this.fechaActualizacion = fechaActualizacion;
 	}
 }
