+++ src/main/java/com/levm/expendienteMedico/entity/IndicacionGeneral.java
@@ -3,7 +3,11 @@ package com.levm.expendienteMedico.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@NoArgsConstructor
 @Entity
 public class IndicacionGeneral {
 	
@@ -13,33 +17,9 @@ public class IndicacionGeneral {
 	
 	private String comentario;
 
-	public IndicacionGeneral() {
-		super();
-	}
-
 	public IndicacionGeneral(int idIndicacion, String comentario) {
 		super();
 		this.idIndicacion = idIndicacion;
 		this.comentario = comentario;
 	}
-
-	public int getIdIndicacion() {
-		return idIndicacion;
-	}
-
-	public void setIdIndicacion(int idIndicacion) {
-		this.idIndicacion = idIndicacion;
-	}
-
-	public String getComentario() {
-		return comentario;
-	}
-
-	public void setComentario(String comentario) {
-		this.comentario = comentario;
-	}
-	
-	
-	
-
 }
