+++ src/main/java/com/levm/expendienteMedico/entity/Medicamento.java
@@ -3,7 +3,11 @@ package com.levm.expendienteMedico.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@NoArgsConstructor
 @Entity
 public class Medicamento {
 	
@@ -15,10 +19,6 @@ public class Medicamento {
 	private String viaAdministracion;
 	private String frecuencia;
 	private String duracion;
-	
-	public Medicamento() {
-		super();
-	}
 
 	public Medicamento(String denominacion, String viaAdministracion, String frecuencia, String duracion) {
 		super();
@@ -28,43 +28,4 @@ public class Medicamento {
 		this.duracion = duracion;
 	}
 
-	public Long getIdMedicamento() {
-		return idMedicamento;
-	}
-
-	public void setIdMedicamento(Long idMedicamento) {
-		this.idMedicamento = idMedicamento;
-	}
-
-	public String getDenominacion() {
-		return denominacion;
-	}
-
-	public void setDenominacion(String denominacion) {
-		this.denominacion = denominacion;
-	}
-
-	public String getViaAdministracion() {
-		return viaAdministracion;
-	}
-
-	public void setViaAdministracion(String viaAdministracion) {
-		this.viaAdministracion = viaAdministracion;
-	}
-
-	public String getFrecuencia() {
-		return frecuencia;
-	}
-
-	public void setFrecuencia(String frecuencia) {
-		this.frecuencia = frecuencia;
-	}
-
-	public String getDuracion() {
-		return duracion;
-	}
-
-	public void setDuracion(String duracion) {
-		this.duracion = duracion;
-	}
 }
