+++ src/main/java/com/levm/expendienteMedico/entity/Receta.java
@@ -3,7 +3,11 @@ package com.levm.expendienteMedico.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@NoArgsConstructor
 @Entity
 public class Receta {
 	
@@ -15,41 +19,9 @@ public class Receta {
 	
 	private boolean alergias;
 
-	public Receta() {
+	public Receta(String padecimiento, boolean alergias) {
 		super();
-	}
-
-	public Receta(int idReceta, String padecimiento, boolean alergias) {
-		super();
-		this.idReceta = idReceta;
 		this.padecimiento = padecimiento;
 		this.alergias = alergias;
 	}
-
-	public int getIdReceta() {
-		return idReceta;
-	}
-
-	public void setIdReceta(int idReceta) {
-		this.idReceta = idReceta;
-	}
-
-	public String getPadecimiento() {
-		return padecimiento;
-	}
-
-	public void setPadecimiento(String padecimiento) {
-		this.padecimiento = padecimiento;
-	}
-
-	public boolean isAlergias() {
-		return alergias;
-	}
-
-	public void setAlergias(boolean alergias) {
-		this.alergias = alergias;
-	}
-	
-	
-
 }
